Rename rascaline.calculator => rascaline.calculators
And add missing documentation for SoapPowerSpectrum
